+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- [@lukehsiao][lh]: Removed the ExternalIPError type.
+
 ### Changed
 - [@lukehsiao][lh]: Switch from `failure` to `thiserror` and `anyhow`.